# backend/src/main/resources/openapi.yaml
openapi: 3.0.4
info:
  title: Sistema de gestión de alojamientos
  description: API para gestionar alojamientos, reservas, usuarios y servicios relacionados.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
  /usuarios:
    get:
      summary: Obtener todos los usuarios
      tags:
        - Usuario
      responses:
        "200":
          description: Lista de usuarios obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UsuarioDTO"
        "400":
          description: Petición inválida
        "404":
          description: No se encontraron usuarios
    post:
      summary: Crear un nuevo usuario
      tags:
        - Usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioDTO"
            example:
              nombre: "David Gómez"
              email: "david@example.com"
              rol: "CLIENTE"
      responses:
        "200":
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsuarioDTO"
        "400":
          description: Datos inválidos
        "404":
          description: Error al crear usuario
...
  /reservas:
    get:
      summary: Obtener todas las reservas
      tags:
        - Reserva
      responses:
        "200":
          description: Lista de reservas obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReservaDTO"
              example:
                - id: 1
                  alojamientoId: 1
                  usuarioId: 2
                  fechaInicio: "2023-07-01"
                  fechaFin: "2023-07-07"
                - id: 2
                  alojamientoId: 2
                  usuarioId: 3
                  fechaInicio: "2023-08-10"
                  fechaFin: "2023-08-15"
        "400":
          description: Error en la solicitud
        "404":
          description: No se encontraron reservas

    post:
      summary: Crear una nueva reserva
      tags:
        - Reserva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservaDTO"
            example:
              alojamientoId: 1
              usuarioId: 2
              fechaInicio: "2023-09-01"
              fechaFin: "2023-09-05"
      responses:
        "201":
          description: Reserva creada con éxito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservaDTO"
              example:
                id: 5
                alojamientoId: 1
                usuarioId: 2
                fechaInicio: "2023-09-01"
                fechaFin: "2023-09-05"
        "400":
          description: Datos inválidos para crear la reserva
        "404":
          description: No se pudo crear la reserva
...
  /servicios:
    get:
      summary: Obtener todos los servicios disponibles
      tags:
        - Servicio
      responses:
        "200":
          description: Lista de servicios obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServicioDTO"
              example:
                - id: 1
                  nombre: "WiFi"
                  descripcion: "Internet inalámbrico de alta velocidad"
                - id: 2
                  nombre: "Piscina"
                  descripcion: "Piscina al aire libre con agua climatizada"
        "400":
          description: Error en la solicitud
        "404":
          description: No se encontraron servicios

    post:
      summary: Crear un nuevo servicio
      tags:
        - Servicio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServicioDTO"
            example:
              nombre: "Desayuno incluido"
              descripcion: "Desayuno buffet disponible todas las mañanas"
      responses:
        "201":
          description: Servicio creado con éxito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServicioDTO"
              example:
                id: 5
                nombre: "Desayuno incluido"
                descripcion: "Desayuno buffet disponible todas las mañanas"
        "400":
          description: Datos inválidos para crear servicio
        "409":
          description: Ya existe un servicio con el mismo nombre
...
components:
  schemas:
    UsuarioDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nombre:
          type: string
          example: David Gómez
        email:
          type: string
          example: david@example.com
        rol:
          type: string
          example: CLIENTE
...
    ReservaDTO:
      type: object
      properties:
        id:
          type: integer
          example: 5
        usuarioId:
          type: integer
          example: 1
        alojamientoId:
          type: integer
          example: 2
        fechaInicio:
          type: string
          format: date
          example: 2025-09-15
        fechaFin:
          type: string
          format: date
          example: 2025-09-20
        estado:
          type: string
          example: CONFIRMADA
...
    ServicioDTO:
      type: object
      properties:
        id:
          type: integer
          example: 3
        nombre:
          type: string
          example: Piscina
        descripcion:
          type: string
          example: Piscina climatizada disponible todo el año
...